ai-analyst/requirements.txt
requests
pandas
anthropic
schedule
python-dotenv
feedparser
redis
